fix:red dot notification placement
